core/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.2
django-cors-headers==4.1.0
django-restframework==0.0.1
djangorestframework==3.14.0
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.5.0
psycopg2-binary==2.9.6
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
uritemplate==4.1.1
whitenoise==6.5.0
